separate builtin witness
