+++ stwo_cairo_prover/crates/prover/src/witness/cairo.rs
@@ -4,6 +4,7 @@ use stwo_prover::core::backend::simd::SimdBackend;
 use stwo_prover::core::fields::m31::M31;
 use tracing::{span, Level};
 
+use super::builtins::{BuiltinsClaimGenerator, BuiltinsInteractionClaimGenerator};
 use super::opcodes::{OpcodesClaimGenerator, OpcodesInteractionClaimGenerator};
 use crate::cairo_air::air::{
     CairoClaim, CairoInteractionClaim, CairoInteractionElements, PublicData,
@@ -11,7 +12,6 @@ use crate::cairo_air::air::{
 use crate::cairo_air::blake::air::{
     BlakeContextClaimGenerator, BlakeContextInteractionClaimGenerator,
 };
-use crate::cairo_air::builtins_air::{BuiltinsClaimGenerator, BuiltinsInteractionClaimGenerator};
 use crate::cairo_air::poseidon::air::{
     PoseidonContextClaimGenerator, PoseidonContextInteractionClaimGenerator,
 };
