+++ stwo_cairo_prover/crates/prover/src/witness/mod.rs
@@ -1,2 +1,3 @@
+pub mod builtins;
 pub mod cairo;
 pub mod opcodes;
